squashed commit to underp settings menu for tablets.  Sorry!!!
So sorry for this major derp for tablet users..

it was recently brought to my attention that tablet kept
getting force closes when trying to access certain menus.
The issue is that i followed CM's commit that setup tablets to use
dual pane, but my custom menus and java files kept the regular
single pane setup.  Obviously it doesnt take a genius to figure
out the outcome.  Revert everything back to single pane for tablets.
Much easier and the way google wants it.

Change-Id: I8b0213b4a8f66d6cb5fc72c9837f267e4341727c
Signed-off-by: scott <scott@ubuntu>